requirements.txt
python-dotenv==1.0.1
streamlit==1.37.0
langchain==0.2.12
langchain_google_genai==1.0.8
langchain_core==0.2.29